# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Social Network API",
    "description": "API documentation",
    "version": "0.1.9"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/posts/{id}": {
      "get": {
        "tags": [
          "Post features"
          ],
        "summary": "Get post by id",
        "description": "Get post by id",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "schema": {
              "type": "integer"
            },
            "description": "post id \n"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Post features"
          ],
        "summary": "update post by id",
        "description": "update post by id",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "schema": {
              "type": "integer"
            },
            "description": "post id \n",
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "caption": {
                    "type": "string"
                  },
                  "urlImage": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/images"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Post features"
          ],
        "summary": "delete post by id",
        "description": "delete post by id",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "schema": {
              "type": "integer"
            },
            "description": "post id \n",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Post features"
          ],
        "summary": "Like or unlike post",
        "description": "Like or unlike post",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "schema": {
              "type": "integer"
            },
            "description": "post id \n"
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "status like or unlike post \n"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "images": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "readOnly": true
          },
          "image_path": {
            "type": "string"
          },
          "downsize": {
            "type": "string"
          },
          "thumbnail": {
            "type": "string"
          },
          "metadata": {
            "type": "string"
          },
          "post_id": {
            "type": "integer",
            "readOnly": true
          }
        },
        "require": [
          "id",
          "image_path",
          "downsize",
          "thumbnail",
          "metadata",
          "post_id"
        ]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "apiKey",
        "description": "Enter your token here",
        "name": "Authorization",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}